feat(management): Allow groups as PO of an API
* Manage API PO in a group
  * add restriction to prevent defining PRIMARY_OWNER as default role in a group
  * authorize at most 1 API PRIMARY OWNER member in a group
  * return the API primary owner of the group. It is maintained up to date on every action on group members.

* Manage po groups in import/export of an api
* Prevent po group to be added to an API
* Allow transfer ownership to a group
* Add a "group only" mode
  * clean code
  * add a new parameter to manage primary owner mode
  * use this parameter to determine the way API creation from definition is done
  * keep only user members in export
  * when a group is po of an API, it is also attached to the API. When this group is no more a PO, it is removed from API groups.

Closes gravitee-io/issues#5133
Closes gravitee-io/issues#5135
Closes gravitee-io/issues#5136
Closes gravitee-io/issues#5137
Closes gravitee-io/issues#5162

+++ gravitee-rest-api-service/src/main/java/io/gravitee/rest/api/service/impl/GroupServiceImpl.java
@@ -30,13 +30,11 @@ import io.gravitee.rest.api.model.api.ApiEntity;
 import io.gravitee.rest.api.model.permissions.RolePermission;
 import io.gravitee.rest.api.model.permissions.RoleScope;
 import io.gravitee.rest.api.model.permissions.SystemRole;
+import io.gravitee.rest.api.model.settings.ApiPrimaryOwnerMode;
 import io.gravitee.rest.api.service.*;
 import io.gravitee.rest.api.service.common.GraviteeContext;
 import io.gravitee.rest.api.service.common.RandomString;
-import io.gravitee.rest.api.service.exceptions.GroupNameAlreadyExistsException;
-import io.gravitee.rest.api.service.exceptions.GroupNotFoundException;
-import io.gravitee.rest.api.service.exceptions.GroupsNotFoundException;
-import io.gravitee.rest.api.service.exceptions.TechnicalManagementException;
+import io.gravitee.rest.api.service.exceptions.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -82,7 +80,7 @@ public class GroupServiceImpl extends AbstractService implements GroupService {
     private PlanRepository planRepository;
     @Autowired
     private IdentityProviderRepository identityProviderRepository;
-    
+
     @Override
     public List<GroupEntity> findAll() {
         try {
@@ -93,6 +91,8 @@ public class GroupServiceImpl extends AbstractService implements GroupService {
                     .map(this::map)
                     .sorted(Comparator.comparing(GroupEntity::getName))
                     .collect(Collectors.toList());
+
+            populateGroupFlags(groups);
 
             if (permissionService.hasPermission(RolePermission.ENVIRONMENT_GROUP, GraviteeContext.getCurrentEnvironment(), CREATE, UPDATE, DELETE)) {
                 groups.forEach(groupEntity -> groupEntity.setManageable(true));
@@ -115,6 +115,19 @@ public class GroupServiceImpl extends AbstractService implements GroupService {
             logger.error("An error occurs while trying to find all groups", ex);
             throw new TechnicalManagementException("An error occurs while trying to find all groups", ex);
         }
+    }
+
+    private void populateGroupFlags(final List<GroupEntity> groups) {
+        RoleEntity apiPORole = roleService.findByScopeAndName(RoleScope.API, SystemRole.PRIMARY_OWNER.name())
+                .orElseThrow(() -> new TechnicalManagementException("API System Role 'PRIMARY_OWNER' not found."));
+
+        groups.forEach(group -> {
+            final boolean isApiPO = !membershipService.getMembershipsByMemberAndReferenceAndRole(
+                    MembershipMemberType.GROUP, group.getId(), MembershipReferenceType.API, apiPORole.getId())
+                    .isEmpty();
+
+            group.setPrimaryOwner(isApiPO);
+        });
     }
 
     @Override
@@ -185,8 +198,8 @@ public class GroupServiceImpl extends AbstractService implements GroupService {
             GroupEntity grp = this.map(groupRepository.update(updatedGroup));
             logger.debug("update {} - DONE", grp);
 
-            updateDefautRoles(groupId, updatedGroupEntity.getRoles(), group.getRoles());
-            
+            updateDefaultRoles(groupId, updatedGroupEntity.getRoles(), group.getRoles());
+
             // Audit
             auditService.createEnvironmentAuditLog(
                     Collections.singletonMap(GROUP, groupId),
@@ -202,19 +215,28 @@ public class GroupServiceImpl extends AbstractService implements GroupService {
         }
     }
 
-    private void updateDefautRoles(String groupId, Map<RoleScope, String> formerRoles, Map<RoleScope, String> newRoles) throws TechnicalException {
+    private void updateDefaultRoles(String groupId, Map<RoleScope, String> formerRoles, Map<RoleScope, String> newRoles) throws TechnicalException {
         RoleScope[] groupRoleScopes = { RoleScope.API, RoleScope.APPLICATION };
-        for(RoleScope roleScope: groupRoleScopes) {
+        for (RoleScope roleScope : groupRoleScopes) {
             if (
-                    (formerRoles != null && formerRoles.get(roleScope) != null) 
-                    && (newRoles == null || (newRoles != null && !formerRoles.get(roleScope).equals(newRoles.get(roleScope))))
-                    ) {
+                    formerRoles != null &&
+                            formerRoles.get(roleScope) != null &&
+                            (newRoles == null ||
+                                    (newRoles != null &&
+                                            !formerRoles.get(roleScope).equals(newRoles.get(roleScope)) &&
+                                            !SystemRole.PRIMARY_OWNER.name().equals(newRoles.get(roleScope))))
+            ) {
                 removeOldDefaultRole(groupId, MembershipReferenceType.valueOf(roleScope.name()));
             }
+
             if (
-                    (newRoles != null && newRoles.get(roleScope) != null) 
-                    && (formerRoles == null || (formerRoles != null && !newRoles.get(roleScope).equals(formerRoles.get(roleScope))))
-                    ) {
+                    newRoles != null &&
+                            newRoles.get(roleScope) != null  &&
+                            !SystemRole.PRIMARY_OWNER.name().equals(newRoles.get(roleScope)) &&
+                            (formerRoles == null ||
+                                    (formerRoles != null &&
+                                            !newRoles.get(roleScope).equals(formerRoles.get(roleScope))))
+            ) {
                 addNewDefaultRole(groupId, newRoles.get(roleScope), roleScope);
             }
         }
@@ -231,7 +253,7 @@ public class GroupServiceImpl extends AbstractService implements GroupService {
                 new MembershipService.MembershipRole(roleScope, newRole));
 
     }
-    
+
     @Override
     public GroupEntity findById(String groupId) {
         try {
@@ -370,9 +392,22 @@ public class GroupServiceImpl extends AbstractService implements GroupService {
             if (!group.isPresent()) {
                 throw new GroupNotFoundException(groupId);
             }
+
+            RoleEntity apiPORole = roleService.findByScopeAndName(RoleScope.API, SystemRole.PRIMARY_OWNER.name())
+                    .orElseThrow(() -> new TechnicalManagementException("API System Role 'PRIMARY_OWNER' not found."));
+
+
+            final long apiCount = membershipService.getMembershipsByMemberAndReferenceAndRole(
+                    MembershipMemberType.GROUP, groupId, MembershipReferenceType.API, apiPORole.getId())
+                    .size();
+
+            if(apiCount > 0) {
+                throw new StillPrimaryOwnerException(apiCount, ApiPrimaryOwnerMode.GROUP);
+            }
+
             //remove all members
             membershipService.deleteReference(MembershipReferenceType.GROUP, groupId);
-            
+
             //remove all applications or apis
             Date updatedDate = new Date();
             apiRepository.search(new ApiCriteria.Builder().environmentId(GraviteeContext.getCurrentEnvironment()).groups(groupId).build()).forEach(api -> {
@@ -506,17 +541,17 @@ public class GroupServiceImpl extends AbstractService implements GroupService {
         }
 
         // in connected mode,
-        
+
         // if no restriction defined
         if (excludedGroups == null || excludedGroups.isEmpty()) {
             return true;
         }
-        
+
         // if user is a direct member of the API
         if (!membershipService.getRoles(MembershipReferenceType.API, api.getId(), MembershipMemberType.USER, username).isEmpty()) {
             return true;
         }
-        
+
         // for public apis
         // user must not be a member of any exclusion group.
         // That is user must not have no API role on each of the exclusion groups
@@ -527,7 +562,7 @@ public class GroupServiceImpl extends AbstractService implements GroupService {
                             .noneMatch(role -> role.getScope() == RoleScope.API)
                     );
         }
-        
+
         // for private apis
         // user must be in at least one attached group which is not also an exclusion group.
         if (Visibility.PRIVATE.equals(api.getVisibility()) && api.getGroups() != null && !api.getGroups().isEmpty()) {
@@ -554,7 +589,7 @@ public class GroupServiceImpl extends AbstractService implements GroupService {
         }
 
         // in connected mode
-        
+
         // if no restriction defined
         if (excludedGroups == null || excludedGroups.isEmpty()) {
             return true;
@@ -684,6 +719,7 @@ public class GroupServiceImpl extends AbstractService implements GroupService {
         GroupEntity entity = new GroupEntity();
         entity.setId(group.getId());
         entity.setName(group.getName());
+        entity.setApiPrimaryOwner(group.getApiPrimaryOwner());
 
         if(group.getEventRules() != null && !group.getEventRules().isEmpty()) {
             List<GroupEventRuleEntity> groupEventRules = new ArrayList<>();
@@ -723,15 +759,10 @@ public class GroupServiceImpl extends AbstractService implements GroupService {
         //check if user exist
         this.userService.findById(username);
         membershipService.deleteReferenceMember(MembershipReferenceType.GROUP, groupId, MembershipMemberType.USER, username);
-    }
 
-    @Override
-    public void addUserToGroup(String groupId, String username, String... roleIds) {
-        //check if user exist
-        this.userService.findById(username);
-        
-        for(String roleId: roleIds) {
-            membershipService.addRoleToMemberOnReference(MembershipReferenceType.GROUP, groupId, MembershipMemberType.USER, username, roleId);
+        GroupEntity existingGroup = this.findById(groupId);
+        if(existingGroup.getApiPrimaryOwner() != null && existingGroup.getApiPrimaryOwner().equals(username)) {
+            updateApiPrimaryOwner(groupId, username);
         }
     }
 
@@ -745,7 +776,24 @@ public class GroupServiceImpl extends AbstractService implements GroupService {
 
     @Override
     public int getNumberOfMembers(String groupId) {
-        return membershipService.getMembersByReference(MembershipReferenceType.GROUP, groupId).size() + 
+        return membershipService.getMembersByReference(MembershipReferenceType.GROUP, groupId).size() +
                 invitationService.findByReference(InvitationReferenceType.GROUP, groupId).size();
     }
+
+    @Override
+    public void updateApiPrimaryOwner(String groupId, String newApiPrimaryOwner) {
+        try {
+            Optional<Group> group = groupRepository.findById(groupId);
+            if (!group.isPresent()) {
+                throw new GroupNotFoundException(groupId);
+            }
+            final Group foundGroup = group.get();
+            foundGroup.setApiPrimaryOwner(newApiPrimaryOwner);
+            groupRepository.update(foundGroup);
+        } catch (TechnicalException ex) {
+            logger.error("An error occurs while trying to find or update a group", ex);
+            throw new TechnicalManagementException("An error occurs while trying to find or update a group", ex);
+        }
+    }
+
 }
